move to double from float, compute acceleration

+++ include/body.h
@@ -7,7 +7,7 @@ typedef struct phys_body {
 	phys_vector2 pos;
 	phys_vector2 vel;
 	phys_vector2 accel;
-	float mass;
+	double mass;
 	phys_vector2 force;
 
 } phys_body;
@@ -17,7 +17,7 @@ typedef struct phys_system {
 	size_t count;
 } phys_system;
 
-void phys_init_body(phys_body* body, phys_vector2 pos, phys_vector2 vel, phys_vector2 accel, float mass);
+void phys_init_body(phys_body* body, phys_vector2 pos, phys_vector2 vel, phys_vector2 accel, double mass);
 void phys_init_system(phys_system* system, phys_body* bodies, size_t count);
 
 #endif
